split into 2 clasess
